# docs/swagger.yaml
basePath: /
definitions:
  audits.AuditCheckResponse:
    properties:
      data:
        $ref: '#/definitions/models.AuditCheck'
    type: object
  audits.AuditChecksResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.AuditCheck'
        type: array
    type: object
  audits.CreateAuditCheckData:
    properties:
      category:
        type: string
      filter_config:
        additionalProperties: {}
        type: object
      name:
        type: string
      search_keyword_url_id:
        type: integer
    type: object
  audits.CreateAuditCheckRequest:
    properties:
      data:
        $ref: '#/definitions/audits.CreateAuditCheckData'
    type: object
  audits.DeleteAuditCheckData:
    properties:
      id:
        type: integer
    type: object
  audits.DeleteAuditCheckResponse:
    properties:
      data:
        $ref: '#/definitions/audits.DeleteAuditCheckData'
    type: object
  audits.UpdateAuditCheckData:
    properties:
      category:
        type: string
      filter_config:
        additionalProperties: {}
        type: object
      name:
        type: string
    type: object
  audits.UpdateAuditCheckRequest:
    properties:
      data:
        $ref: '#/definitions/audits.UpdateAuditCheckData'
      id:
        type: integer
    type: object
...
  models.AuditCheck:
    properties:
      category:
        type: string
      createdAt:
        type: string
      filterConfig:
        additionalProperties: {}
        type: object
      id:
        type: integer
      name:
        type: string
      searchKeywordURLID:
        type: integer
      updatedAt:
        type: string
    type: object
...
  models.Page:
    properties:
      crawling_session_id:
        type: integer
      id:
        type: integer
      response_code:
        type: integer
      url:
        type: string
    type: object
  models.PageImage:
    properties:
      id:
        type: integer
      page_id:
        type: integer
      url:
        type: string
    type: object
  models.View:
    properties:
      createdAt:
        type: string
      filterConfig:
        additionalProperties: {}
        type: object
      id:
        type: integer
      name:
        type: string
      searchKeywordURLID:
        type: integer
      updatedAt:
        type: string
    type: object
...
  stats.PageDetailsData:
    properties:
      broken_pages:
        items:
          $ref: '#/definitions/models.Page'
        type: array
      page_images:
        items:
          $ref: '#/definitions/models.PageImage'
        type: array
      pages_linked_to_broken_page:
        items:
          $ref: '#/definitions/models.Page'
        type: array
    type: object
  stats.PageDetailsResponse:
    properties:
      data:
        $ref: '#/definitions/stats.PageDetailsData'
    type: object
...
  views.CreateViewData:
    properties:
      filter_config:
        additionalProperties: {}
        type: object
      name:
        type: string
      search_keyword_url_id:
        type: integer
    type: object
  views.CreateViewRequest:
    properties:
      data:
        $ref: '#/definitions/views.CreateViewData'
    type: object
...
  views.ViewResponse:
    properties:
      data:
        $ref: '#/definitions/models.View'
    type: object
  views.ViewsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.View'
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Access token for authentication
  termsOfService: http://swagger.io/terms/
  title: SiteCrawler API
  version: "1.0"
paths:
  /api/audit_checks:
    get:
      description: Lists audit checks for a given `search_keyword_url_id`
      parameters:
      - description: Search keyword URL ID
        in: query
        name: search_keyword_url_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/audits.AuditChecksResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List audit checks
      tags:
      - AuditChecks
    post:
      consumes:
      - application/json
      description: Creates a new audit check for a given `search_keyword_url_id`
      parameters:
      - description: Audit check payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/audits.CreateAuditCheckRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/audits.AuditCheckResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create audit check
      tags:
      - AuditChecks
  /api/audit_checks/{id}:
    delete:
      description: Deletes an audit check by ID
      parameters:
      - description: Audit check ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/audits.DeleteAuditCheckResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete audit check
      tags:
      - AuditChecks
    get:
      description: Fetches an audit check by ID
      parameters:
      - description: Audit check ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/audits.AuditCheckResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get audit check
      tags:
      - AuditChecks
    put:
      consumes:
      - application/json
      description: Updates an audit check by ID
      parameters:
      - description: Audit check ID
        in: path
        name: id
        required: true
        type: integer
      - description: Audit check update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/audits.UpdateAuditCheckRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/audits.AuditCheckResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update audit check
      tags:
      - AuditChecks
...
  /api/pages/{id}/page_details:
    get:
      description: Returns images/referrers for a page depending on its status
      parameters:
      - description: Page ID
        in: path
        name: id
        required: true
        type: integer
      - description: Search keyword URL ID
        in: query
        name: search_keyword_url_id
        required: true
        type: integer
      - description: Result limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/stats.PageDetailsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get page details
      tags:
      - Pages
...
  /api/views:
    get:
      description: Lists views for a given `search_keyword_url_id`
      parameters:
      - description: Search keyword URL ID
        in: query
        name: search_keyword_url_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.ViewsResponse'
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List views
      tags:
      - Views
    post:
      consumes:
      - application/json
      description: Creates a new view
      parameters:
      - description: View payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/views.CreateViewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/views.ViewResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create view
      tags:
      - Views
...
securityDefinitions:
  ApiKeyAuth:
    description: Access token for authentication
    in: header
    name: X-Access-Token
    type: apiKey
swagger: "2.0"